requirements.txt
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cs50==7.0.2
Flask==2.0.2
Flask-Session==0.4.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
python-decouple==3.5
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.7
Werkzeug==2.0.2
